+++ src/@next/components/atoms/ErrorMessage/stories.tsx
@@ -1,11 +1,11 @@
+import { storiesOf } from "@storybook/react";
 import React from "react";
 
 import { ErrorMessage } from ".";
-import { createStory } from "../baseStory";
 
 const ERRORS = [{ field: "Field", message: "Error Message" }];
 
-createStory("ErrorMessage")
+storiesOf("@components/atoms/ErrorMessage", module)
   .add("default", () => <ErrorMessage errors={ERRORS} />)
   .add("with multiple errors", () => (
     <ErrorMessage errors={[...ERRORS, ...ERRORS]} />