+++ src/@next/components/atoms/Icon/stories.tsx
@@ -1,6 +1,8 @@
+import { storiesOf } from "@storybook/react";
 import React from "react";
 
 import { Icon } from ".";
-import { createStory } from "../baseStory";
 
-createStory("Icon").add("sample icon", () => <Icon name="arrow_back" />);
+storiesOf("@components/atoms/Icon", module).add("sample icon", () => (
+  <Icon name="arrow_back" />
+));