+++ src/@next/components/atoms/Input/stories.tsx
@@ -1,10 +1,10 @@
 import { text } from "@storybook/addon-knobs";
+import { storiesOf } from "@storybook/react";
 import React from "react";
 
 import { Input } from ".";
-import { createStory } from "../baseStory";
 
-createStory("Input")
+storiesOf("@components/atoms/Input", module)
   .add("default", () => <Input value="" label="Empty" />)
   .add("with value", () => (
     <Input