+++ src/@next/components/atoms/RichTextContent/stories.tsx
@@ -1,11 +1,11 @@
+import { storiesOf } from "@storybook/react";
 import React from "react";
 
 import { RichTextContent } from ".";
-import { createStory } from "../baseStory";
 import descriptionJson from "./fixtures/default_text_block";
 import customDescriptionJson from "./fixtures/text_blocks";
 
-createStory("RichTextContent")
+storiesOf("@components/atoms/RichTextContent", module)
   .add("default", () => <RichTextContent descriptionJson={descriptionJson} />)
   .add("custom", () => (
     <RichTextContent descriptionJson={customDescriptionJson} />