+++ src/@next/components/atoms/SocialMediaIcon/stories.tsx
@@ -1,7 +1,7 @@
+import { storiesOf } from "@storybook/react";
 import React from "react";
 
 import { SocialMediaIcon } from ".";
-import { createStory } from "../baseStory";
 import { Medium } from "./types";
 
 const FACEBOOK_MEDIUM: Medium = {
@@ -10,6 +10,9 @@ const FACEBOOK_MEDIUM: Medium = {
   iconName: "social_facebook",
 };
 
-createStory("SocialMediaIcon").add("sample medium", () => (
-  <SocialMediaIcon medium={FACEBOOK_MEDIUM} key={FACEBOOK_MEDIUM.ariaLabel} />
-));
+storiesOf("@components/atoms/SocialMediaIcon", module).add(
+  "sample medium",
+  () => (
+    <SocialMediaIcon medium={FACEBOOK_MEDIUM} key={FACEBOOK_MEDIUM.ariaLabel} />
+  )
+);