+++ src/@next/components/molecules/baseStory.ts
@@ -1,4 +1,0 @@
-import { storiesOf } from "@storybook/react";
-
-export const createStory = (name: string = "default") =>
-  storiesOf(`@components/molecules/${name}`, module);